refactor(core): vectorize physics utilites and improve numerical stability
- Replace safe_len with native .shape for vectorization
- Enforce float64 precision in JAX backends
- Increase minimum dt threshold to 1e-5 to prevent 0^-beta singularities
- Symmetrize diagonal correction side-bands

+++ homodyne/core/physics_nlsq.py
@@ -218,7 +218,7 @@ def _compute_g1_shear_meshgrid(
 
     # Fix phi shape if it has extra dimensions
     # Handle case where phi might be (1, 1, 1, 23) instead of (23,) or other malformed shapes
-    phi = jnp.asarray(phi)  # Ensure it's a JAX array
+    phi = jnp.asarray(phi, dtype=jnp.float64)  # Ensure it's a JAX float64 array
 
     # Remove all leading singleton dimensions and flatten to 1D
     while phi.ndim > 1:
